Moved File Locations to Improve Cypress Run
