+++ cypress/redundant/home_page_spec.js
