+++ cypress/redundant/user_can_sign_up_spec.js
